+++ src/report/utility/populateDb/populateEndpoints.ts
@@ -4,6 +4,9 @@ import Database from "better-sqlite3";
 const populateEndpoints = async (db: Database.Database, endpoints: any) => {
     const insert = db.prepare("INSERT OR IGNORE INTO endpoints (url) VALUES (?)");
 
+    // clear the endpoints table
+    db.prepare("DELETE FROM endpoints").run();
+
     const insertPaths = (base: string, paths: object) => {
         for (const path in paths) {
             // The path is the key itself, which is a full path